grading

+++ Source/AudioAnalyzer.h
@@ -49,7 +49,7 @@ public:
     int FrameAnalysis(const float *buffer, float *out);
     int SubbandAnalysis(std::vector<float> &subband_signal, uint32_t midi_note); // find notes
     int SetScore(struct Note *score, float *times, uint32_t note_num);
-    int AudioScoreAlignment(/*std::vector<std::pair<TimedNotes::iterator, TimedNotes::iterator>> &alignment*/);
+    float AudioScoreAlignment(/*std::vector<std::pair<TimedNotes::iterator, TimedNotes::iterator>> &alignment*/);
     int Clear();
     float fs;
     uint32_t frame_size;
